requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
supabase==2.10.0
pydantic==2.10.5
python-dotenv==1.0.0
web3==6.15.1
eth-typing==3.5.0
httpx==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
faker==22.2.0
sentry-sdk[fastapi]==2.9.0
prometheus-client==0.19.0
celery==5.3.4
redis==5.0.1
a2wsgi==1.10.8
